src/requirements.txt
Flask
flask-cors
argon2-cffi
Flask-SQLAlchemy
pyjwt
pytest